Extract SessionCompletionController from IssueTracker (CROW-1094)
Move auto-complete + auto-cleanup (syncInReviewSessions,
autoCompleteFinishedSessions/Reviews, autoCleanupExpiredSessions) and the pure
decision helpers (decideSessionCompletions, decideReviewCompletions,
sessionsEligibleForCleanup, reviewKickoffAction) plus CompletionDecision/
CompletionResult/ReviewKickoffAction into their own collaborator. It reads
appState and fires its completion callbacks through an unowned back-reference;
refresh() drives it in the same order.

The class is public so the cross-module review-kickoff decision keeps its home;
an `extension IssueTracker` re-exposes the exact `IssueTracker.<symbol>`
spelling that ReviewsPayload, SessionService, and the tests already use, so no
call site or test assertion changes.

swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
